+++ .storybook/preview.tsx
@@ -53,18 +53,18 @@ export const decorators: Decorator[] = [
           <Story />
         </div>
         <div // left column
-          className="col-span-1 col-start-1 row-start-1 hidden bg-gradient-to-r
-            from-slate-400 via-slate-500 to-slate-700 dark:from-slate-300 
-            dark:via-slate-400 dark:to-slate-500 sm:block"
+          className="col-span-1 col-start-1 row-span-3 row-start-1 hidden bg-gradient-to-r
+          from-slate-300 via-slate-400 to-slate-500 dark:from-slate-800
+          dark:via-slate-700 dark:to-slate-600 sm:block"
         >
-          <div className="invisible h-full w-full bg-opacity-5 bg-gradient-to-r from-rose-50 to-transparent dark:visible" />
+          <div className="invisible h-full w-full bg-gradient-to-l from-rose-50 to-slate-700 opacity-25 dark:visible" />
         </div>
         <div // right column
-          className="col-span-1 col-start-3 row-start-1 hidden bg-gradient-to-l
-            from-slate-400 via-slate-500 to-slate-700 dark:from-slate-300 
-            dark:via-slate-400 dark:to-slate-500 sm:block"
+          className="col-span-1 col-start-3 row-span-3 row-start-1 hidden bg-gradient-to-l
+          from-slate-300 via-slate-400 to-slate-500 dark:from-slate-800 
+          dark:via-slate-700 dark:to-slate-600 sm:block"
         >
-          <div className="invisible h-full w-full bg-opacity-5 bg-gradient-to-l from-rose-50 to-transparent dark:visible" />
+          <div className="invisible h-full w-full bg-gradient-to-r from-rose-50 to-slate-700 opacity-25 dark:visible" />
         </div>
       </div>
     </div>
