+++ lib/search.test.ts
@@ -1,7 +1,7 @@
 import { searchPosts, getTagsWithCount } from './search';
-import { getDummyPosts } from '@/lib/mdx/mocks';
+import { getDummyNodes } from '@/lib/mdx/mocks';
 
-const posts = getDummyPosts(10, (index) => ({
+const posts = getDummyNodes(10, (index) => ({
   title: `Post ${index + 1}`,
   tags: ['tag', `tag${index + 1}`],
   excerpt: `Post ${index + 1} excerpt`,
