+++ lib/search.ts
@@ -1,20 +1,16 @@
 import { type BlogMdxNode } from '@/lib/mdx/sources';
 
 /**
- * Search for a query in a text
- * @param query The query to search for
- * @param text The text to search in
- * @returns A boolean indicating whether the query was found in the text
+ * Search for a query in a text.
+ * @returns A boolean indicating whether the query was found in the text.
  */
 function searchHit(query: string, text: string) {
   return text.toLowerCase().includes(query.toLowerCase());
 }
 
 /**
- * Search for a query in a list of posts
- * @param query The query to search for
- * @param posts The posts to search in
- * @returns The posts that matched the query in descending order of relevance
+ * Search for a query in a list of posts.
+ * @returns The posts that matched the query in descending order of relevance.
  */
 export function searchPosts(query: string, posts: Array<BlogMdxNode>) {
   const postsWithSearchHits = new Map<BlogMdxNode, number>();
@@ -54,16 +50,13 @@ export function searchPosts(query: string, posts: Array<BlogMdxNode>) {
 
 /**
  * Get all tags with their count
- * @param posts The posts to get the tags from
  * @returns The tags with their count in descending order of count
  */
 export function getTagsWithCount(posts: Array<BlogMdxNode>) {
   const tagsWithCount = new Map<string, number>();
 
   posts.forEach((post) => {
-    const tags = post.frontmatter.tags;
-
-    tags.forEach((tag) => {
+    post.frontmatter.tags.forEach((tag) => {
       tagsWithCount.set(tag, (tagsWithCount.get(tag) ?? 0) + 1);
     });
   });
